Fixed weird bug with ros package version
