--- examples/offline_execution/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/offline_execution
-[install]
-install_scripts=$base/lib/offline_execution
