+++ offline_executor/offline_executor/execute.py
@@ -3,13 +3,13 @@ from tree_translator import factory
 from ament_index_python.packages import get_package_share_directory
 import os
 
-def execute_tree():
+def execute_main():
 
     # Init ros
     rclpy.init(args=None)
 
     # Get the path to the root of the package
-    demo_root_dir = get_package_share_directory('offline_execution')
+    demo_root_dir = get_package_share_directory('offline_executor')
 
     # Now, you can build paths relative to the root of the package
     tree_file_path = os.path.join(demo_root_dir, 'resource', 'final_tree.xml')
