+++ offline_executor/offline_executor/parse_xml.py
@@ -5,8 +5,8 @@ from ament_index_python.packages import get_package_share_directory, get_package
 def main():
 
     # Get the path to the root of the package
-    demo_share_dir = get_package_share_directory('offline_execution')
-    demo_root_dir = get_package_prefix('offline_execution')
+    demo_share_dir = get_package_share_directory('offline_executor')
+    demo_root_dir = get_package_prefix('offline_executor')
 
     # Init tree and action paths
     init_tree_path = os.path.join(demo_share_dir, 'resource', 'basic_tree.xml')
@@ -19,7 +19,7 @@ def main():
     result_tree_path = os.path.join(demo_share_dir, 'resource', 'final_tree.xml')
 
     # Adjust the path to point to the src/resource directory
-    src_resource_dir = os.path.join(demo_root_dir, '..', '..', 'src', 'offline_execution', 'resource')
+    src_resource_dir = os.path.join(demo_root_dir, '..', '..', 'src', 'offline_executor', 'resource')
     readable_tree_path = os.path.join(src_resource_dir, 'final_tree.xml')
 
     # Store the string in a temp xml file
