+++ offline_executor/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/offline_executor
+[install]
+install_scripts=$base/lib/offline_executor
